refactor: 컴포넌트 네이밍 컨벤션 통일 (#28) - kebab-case => PascalCase 변경
