+++ src/components/fields/index.ts
@@ -1,4 +1,5 @@
-export * from "./date-field";
-export * from "./password-field";
-export * from "./file-field";
-export * from "./select-field";
+export * from "./DateField";
+export * from "./PasswordField";
+export * from "./FileField";
+export * from "./SelectField";
+
